Avoiding pub(crate) persistent_storage
