+++ src/many-ledger/src/storage.rs
@@ -1,14 +1,16 @@
 use crate::error;
 #[cfg(feature = "migrate_blocks")]
 use crate::migration;
-use crate::migration::data::DataMethods;
+use crate::migration::data::{
+    DataMethods, ACCOUNT_TOTAL_COUNT_INDEX, NON_ZERO_ACCOUNT_TOTAL_COUNT_INDEX,
+};
 use crate::migration::{run_migrations, MigrationMap, MigrationSet};
 use crate::module::validate_account;
 use many_error::ManyError;
 use many_identity::Address;
 use many_modules::abci_backend::AbciCommitInfo;
 use many_modules::account::features::FeatureInfo;
-use many_modules::data::{DataIndex, DataInfo};
+use many_modules::data::{DataIndex, DataInfo, DataValue};
 use many_modules::{account, events, idstore, EmptyReturn};
 use many_protocol::ResponseMessage;
 use many_types::ledger::{Symbol, TokenAmount};
@@ -259,7 +261,7 @@ pub(super) fn key_for_multisig_transaction(token: &[u8]) -> Vec<u8> {
 
 pub struct LedgerStorage {
     symbols: BTreeMap<Symbol, String>,
-    pub(crate) persistent_store: merk::Merk,
+    persistent_store: merk::Merk,
 
     /// When this is true, we do not commit every transactions as they come,
     /// but wait for a `commit` call before committing the batch to the
@@ -1561,5 +1563,63 @@ pub mod tests {
             ))
             .len()
         )
+    }
+}
+
+impl DataMethods for LedgerStorage {
+    fn data_attributes(&self) -> Option<BTreeMap<DataIndex, DataValue>> {
+        self.persistent_store
+            .get(DATA_ATTRIBUTES_KEY)
+            .expect("Error while reading the DB")
+            .map(|x| minicbor::decode(&x).unwrap())
+    }
+
+    fn update_data_attributes(
+        &mut self,
+        from: &Address,
+        to: &Address,
+        amount: TokenAmount,
+        symbol: &Address,
+    ) {
+        if let Some(mut attributes) = self.data_attributes() {
+            let key_to = key_for_account_balance(to, symbol);
+            if self
+                .persistent_store
+                .get(&key_to)
+                .expect("Error communicating with the DB")
+                .is_none()
+            {
+                attributes
+                    .entry(*ACCOUNT_TOTAL_COUNT_INDEX)
+                    .and_modify(|x| {
+                        if let DataValue::Counter(count) = x {
+                            *count += 1;
+                        }
+                    });
+                attributes
+                    .entry(*NON_ZERO_ACCOUNT_TOTAL_COUNT_INDEX)
+                    .and_modify(|x| {
+                        if let DataValue::Counter(count) = x {
+                            *count += 1;
+                        }
+                    });
+            }
+            let balance_from = self.get_balance(from, symbol);
+            if balance_from == amount {
+                attributes
+                    .entry(*NON_ZERO_ACCOUNT_TOTAL_COUNT_INDEX)
+                    .and_modify(|x| {
+                        if let DataValue::Counter(count) = x {
+                            *count -= 1;
+                        }
+                    });
+            }
+            self.persistent_store
+                .apply(&[(
+                    DATA_ATTRIBUTES_KEY.to_vec(),
+                    Op::Put(minicbor::to_vec(attributes).unwrap()),
+                )])
+                .unwrap();
+        }
     }
 }
